Extract setting cards
